+++ src/core/engine.rs
@@ -121,10 +121,10 @@ impl<S: AbeScheme> Engine<S> {
         let random = self.random_message()?;
         let enc_sym_key = self
             .encrypt(policy, public_key, attrs, &random)?
-            .as_bytes()?;
+            .try_into_bytes()?;
         // Use a hash of the plaintext bytes as the symmetric key
         let sym_key = Shake256::default()
-            .chain(&random.as_bytes()?)
+            .chain(&random.try_into_bytes()?)
             .finalize_xof()
             .read_boxed(symmetric_key_len)
             .into_vec();
@@ -140,13 +140,13 @@ impl<S: AbeScheme> Engine<S> {
     ) -> Result<Vec<u8>, FormatErr> {
         let random = self
             .decrypt(
-                &S::CipherText::from_bytes(encrypted_symmetric_key)?,
+                &S::CipherText::try_from_bytes(encrypted_symmetric_key)?,
                 decryption_key,
             )?
             .ok_or(FormatErr::InvalidEncryptedData)?;
         // Use a hash of the plaintext bytes as the symmetric key
         Ok(Shake256::default()
-            .chain(&random.as_bytes()?)
+            .chain(&random.try_into_bytes()?)
             .finalize_xof()
             .read_boxed(symmetric_key_len)
             .into_vec())