+++ src/interfaces/ffi/hybrid_gpsw_aes.rs
@@ -94,7 +94,7 @@ pub unsafe extern "C" fn h_aes_create_encryption_cache(
     // Public Key
     let public_key_bytes =
         std::slice::from_raw_parts(public_key_ptr as *const u8, public_key_len as usize);
-    let public_key = match PublicKey::from_bytes(public_key_bytes) {
+    let public_key = match PublicKey::try_from_bytes(public_key_bytes) {
         Ok(key) => key,
         Err(e) => {
             ffi_bail!(format!("Hybrid Cipher: invalid public key: {:?}", e));
@@ -296,7 +296,7 @@ pub unsafe extern "C" fn h_aes_encrypt_header(
     // Public Key
     let public_key_bytes =
         std::slice::from_raw_parts(public_key_ptr as *const u8, public_key_len as usize);
-    let public_key = ffi_unwrap!(PublicKey::from_bytes(public_key_bytes));
+    let public_key = ffi_unwrap!(PublicKey::try_from_bytes(public_key_bytes));
 
     // Attributes
     let attributes = match CStr::from_ptr(attributes_ptr).to_str() {
@@ -416,7 +416,7 @@ pub unsafe extern "C" fn h_aes_create_decryption_cache(
         user_decryption_key_ptr as *const u8,
         user_decryption_key_len as usize,
     );
-    let user_decryption_key = match UserDecryptionKey::from_bytes(user_decryption_key_bytes) {
+    let user_decryption_key = match UserDecryptionKey::try_from_bytes(user_decryption_key_bytes) {
         Ok(key) => key,
         Err(e) => {
             ffi_bail!(format!(
@@ -635,7 +635,8 @@ pub unsafe extern "C" fn h_aes_decrypt_header(
         user_decryption_key_ptr as *const u8,
         user_decryption_key_len as usize,
     );
-    let user_decryption_key = ffi_unwrap!(UserDecryptionKey::from_bytes(user_decryption_key_bytes));
+    let user_decryption_key =
+        ffi_unwrap!(UserDecryptionKey::try_from_bytes(user_decryption_key_bytes));
 
     let header: ClearTextHeader<Aes256GcmCrypto> =
         ffi_unwrap!(decrypt_hybrid_header::<Gpsw<Bls12_381>, Aes256GcmCrypto>(