+++ src/interfaces/wasm_bindgen/hybrid_gpsw_aes_encryption.rs
@@ -62,7 +62,7 @@ pub fn webassembly_encrypt_hybrid_header(
     // Convert JS type
     let policy: Policy = serde_json::from_slice(policy_bytes.to_vec().as_slice())
         .map_err(|e| JsValue::from_str(&format!("Error deserializing Policy: {e:?}")))?;
-    let public_key = PublicKey::from_bytes(public_key_bytes.to_vec().as_slice())
+    let public_key = PublicKey::try_from_bytes(public_key_bytes.to_vec().as_slice())
         .map_err(|e| JsValue::from_str(&format!("Error deserializing Public Key: {e:?}")))?;
 
     let encrypted_header = encrypt_hybrid_header::<Gpsw<Bls12_381>, Aes256GcmCrypto>(
@@ -81,7 +81,7 @@ pub fn webassembly_encrypt_hybrid_header(
     //
     // Flatten struct Encrypted Header
     let encrypted_header_bytes = encrypted_header
-        .as_bytes()
+        .try_into_bytes()
         .map_err(|e| JsValue::from_str(&format!("Error serializing encrypted header: {e:?}")))?;
     Ok(js_sys::Uint8Array::from(&encrypted_header_bytes[..]))
 }
@@ -124,7 +124,7 @@ pub fn webassembly_create_encryption_cache(
 
     //
     // Parse public key
-    let public_key = PublicKey::from_bytes(public_key.to_vec().as_slice()).map_err(|e| {
+    let public_key = PublicKey::try_from_bytes(public_key.to_vec().as_slice()).map_err(|e| {
         return JsValue::from_str(&format!("Error deserializing public key: {e}"));
     })?;
 
@@ -178,7 +178,7 @@ pub fn webassembly_encrypt_hybrid_header_using_cache(
     //
     // Flatten struct Encrypted Header
     let encrypted_header_bytes = encrypted_header
-        .as_bytes()
+        .try_into_bytes()
         .map_err(|e| JsValue::from_str(&format!("Error serializing encrypted header: {e:?}")))?;
     Ok(js_sys::Uint8Array::from(&encrypted_header_bytes[..]))
 }